fix type of length in fft window function

+++ src/include/ngspice/fftext.h
@@ -20,7 +20,7 @@ void fftFree(void);
 // release storage for all private cosine and bit reversed tables
 
 int
-fft_windows(char *window, double *win, double *time, double length, double maxt, double span, int order);
+fft_windows(char *window, double *win, double *time, int length, double maxt, double span, int order);
 // computes some popular window functions
 
 void ffts(double *data, int M, int Rows);
